pyproject.toml
[project]
name = "django-storages-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.2",
    "django-storages>=1.14.4",
    "djangorestframework>=3.15.2",
    "pillow>=10.4.0",
    "boto3>=1.35.39",
    "python-decouple>=3.8",
]
